mobile_operator_app/requirements.txt
streamlit
pandas
psycopg2-binary
pymongo
redis
python-dotenv
pydantic
pytest
pytest-mock
pytest-order